Per-command dependency injection
