+++ src/Command/Acsf/AcsfCommandFactory.php
@@ -12,7 +12,6 @@ use Acquia\Cli\DataStore\CloudDataStore;
 use Acquia\Cli\Helpers\LocalMachineHelper;
 use Acquia\Cli\Helpers\SshHelper;
 use Acquia\Cli\Helpers\TelemetryHelper;
-use GuzzleHttp\Client;
 use Psr\Log\LoggerInterface;
 
 class AcsfCommandFactory implements CommandFactoryInterface {
@@ -28,7 +27,6 @@ class AcsfCommandFactory implements CommandFactoryInterface {
     private SshHelper $sshHelper,
     private string $sshDir,
     private LoggerInterface $logger,
-    private Client $httpClient
   ) {
   }
 
@@ -44,7 +42,6 @@ class AcsfCommandFactory implements CommandFactoryInterface {
       $this->sshHelper,
       $this->sshDir,
       $this->logger,
-      $this->httpClient
     );
   }
 
@@ -60,7 +57,6 @@ class AcsfCommandFactory implements CommandFactoryInterface {
       $this->sshHelper,
       $this->sshDir,
       $this->logger,
-      $this->httpClient
     );
   }
