+++ src/Command/Pull/PullCommandBase.php
@@ -4,25 +4,30 @@ declare(strict_types = 1);
 
 namespace Acquia\Cli\Command\Pull;
 
+use Acquia\Cli\ApiCredentialsInterface;
+use Acquia\Cli\CloudApi\ClientService;
 use Acquia\Cli\Command\CommandBase;
+use Acquia\Cli\DataStore\AcquiaCliDatastore;
+use Acquia\Cli\DataStore\CloudDataStore;
 use Acquia\Cli\Exception\AcquiaCliException;
 use Acquia\Cli\Helpers\IdeCommandTrait;
+use Acquia\Cli\Helpers\LocalMachineHelper;
+use Acquia\Cli\Helpers\SshHelper;
+use Acquia\Cli\Helpers\TelemetryHelper;
 use Acquia\Cli\Output\Checklist;
 use Acquia\DrupalEnvironmentDetector\AcquiaDrupalEnvironmentDetector;
 use AcquiaCloudApi\Connector\Client;
 use AcquiaCloudApi\Endpoints\DatabaseBackups;
-use AcquiaCloudApi\Endpoints\Databases;
 use AcquiaCloudApi\Endpoints\Domains;
-use AcquiaCloudApi\Endpoints\Environments;
 use AcquiaCloudApi\Response\BackupResponse;
 use AcquiaCloudApi\Response\DatabaseResponse;
-use AcquiaCloudApi\Response\DatabasesResponse;
 use AcquiaCloudApi\Response\EnvironmentResponse;
 use Closure;
 use Exception;
 use GuzzleHttp\Exception\RequestException;
 use GuzzleHttp\TransferStats;
 use Psr\Http\Message\UriInterface;
+use Psr\Log\LoggerInterface;
 use React\EventLoop\Loop;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Helper\ProgressBar;
@@ -31,7 +36,6 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\ConsoleOutput;
 use Symfony\Component\Console\Output\ConsoleSectionOutput;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Question\ChoiceQuestion;
 use Symfony\Component\Filesystem\Path;
 
 abstract class PullCommandBase extends CommandBase {
@@ -40,11 +44,25 @@ abstract class PullCommandBase extends CommandBase {
 
   protected Checklist $checklist;
 
-  protected EnvironmentResponse $sourceEnvironment;
-
   private string $site;
 
   private UriInterface $backupDownloadUrl;
+
+  public function __construct(
+    public LocalMachineHelper $localMachineHelper,
+    protected CloudDataStore $datastoreCloud,
+    protected AcquiaCliDatastore $datastoreAcli,
+    protected ApiCredentialsInterface $cloudCredentials,
+    protected TelemetryHelper $telemetryHelper,
+    protected string $projectDir,
+    protected ClientService $cloudApiClientService,
+    public SshHelper $sshHelper,
+    protected string $sshDir,
+    LoggerInterface $logger,
+    protected \GuzzleHttp\Client $httpClient
+  ) {
+    parent::__construct($this->localMachineHelper, $this->datastoreCloud, $this->datastoreAcli, $this->cloudCredentials, $this->telemetryHelper, $this->projectDir, $this->cloudApiClientService, $this->sshHelper, $this->sshDir, $logger);
+  }
 
   /**
    * @see https://github.com/drush-ops/drush/blob/c21a5a24a295cc0513bfdecead6f87f1a2cf91a2/src/Sql/SqlMysql.php#L168
@@ -68,18 +86,6 @@ abstract class PullCommandBase extends CommandBase {
       $table = "`$table`";
     }
     return $tables;
-  }
-
-  protected function getCloudFilesDir(EnvironmentResponse $chosenEnvironment, string $site): string {
-    $sitegroup = self::getSiteGroupFromSshUrl($chosenEnvironment->sshUrl);
-    if ($this->isAcsfEnv($chosenEnvironment)) {
-      return '/mnt/files/' . $sitegroup . '.' . $chosenEnvironment->name . '/sites/g/files/' . $site . '/files';
-    }
-    return $this->getCloudSitesPath($chosenEnvironment, $sitegroup) . "/$site/files";
-  }
-
-  protected function getLocalFilesDir(string $site): string {
-    return $this->dir . '/docroot/sites/' . $site . '/files';
   }
 
   public static function getBackupPath(object $environment, DatabaseResponse $database, object $backupResponse): string {
@@ -110,11 +116,7 @@ abstract class PullCommandBase extends CommandBase {
     return Command::SUCCESS;
   }
 
-  protected function pullCode(InputInterface $input, OutputInterface $output): void {
-    $this->setDirAndRequireProjectCwd($input);
-    $clone = $this->determineCloneProject($output);
-    $sourceEnvironment = $this->determineEnvironment($input, $output, TRUE);
-
+  protected function pullCode(InputInterface $input, OutputInterface $output, bool $clone, EnvironmentResponse $sourceEnvironment): void {
     if ($clone) {
       $this->checklist->addItem('Cloning git repository from the Cloud Platform');
       $this->cloneFromCloud($sourceEnvironment, $this->getOutputCallback($output, $this->checklist));
@@ -167,9 +169,7 @@ abstract class PullCommandBase extends CommandBase {
     }
   }
 
-  protected function pullFiles(InputInterface $input, OutputInterface $output): void {
-    $this->setDirAndRequireProjectCwd($input);
-    $sourceEnvironment = $this->determineEnvironment($input, $output, TRUE);
+  protected function pullFiles(InputInterface $input, OutputInterface $output, EnvironmentResponse $sourceEnvironment): void {
     $this->checklist->addItem('Copying Drupal\'s public files from the Cloud Platform');
     $site = $this->determineSite($sourceEnvironment, $input);
     $this->rsyncFilesFromCloud($sourceEnvironment, $this->getOutputCallback($output, $this->checklist), $site);
@@ -421,129 +421,6 @@ abstract class PullCommandBase extends CommandBase {
     }
   }
 
-  protected function isLocalGitRepoDirty(): bool {
-    $this->localMachineHelper->checkRequiredBinariesExist(['git']);
-    $process = $this->localMachineHelper->executeFromCmd(
-      // Problem with this is that it stages changes for the user. They may
-      // not want that.
-      'git add . && git diff-index --cached --quiet HEAD',
-     NULL, $this->dir, FALSE);
-
-    return !$process->isSuccessful();
-  }
-
-  protected function getLocalGitCommitHash(): string {
-    $this->localMachineHelper->checkRequiredBinariesExist(['git']);
-    $process = $this->localMachineHelper->execute([
-      'git',
-      'rev-parse',
-      'HEAD',
-    ], NULL, $this->dir, FALSE);
-
-    if (!$process->isSuccessful()) {
-      throw new AcquiaCliException("Unable to determine Git commit hash.");
-    }
-
-    return trim($process->getOutput());
-  }
-
-  private function promptChooseEnvironment(Client $acquiaCloudClient, string $applicationUuid, bool $allowProduction = FALSE): EnvironmentResponse {
-    $environmentResource = new Environments($acquiaCloudClient);
-    $applicationEnvironments = iterator_to_array($environmentResource->getAll($applicationUuid));
-    $choices = [];
-    foreach ($applicationEnvironments as $key => $environment) {
-      if (!$allowProduction && $environment->flags->production) {
-        unset($applicationEnvironments[$key]);
-        // Re-index array so keys match those in $choices.
-        $applicationEnvironments = array_values($applicationEnvironments);
-        continue;
-      }
-      $choices[] = "$environment->label, $environment->name (vcs: {$environment->vcs->path})";
-    }
-    $chosenEnvironmentLabel = $this->io->choice('Choose a Cloud Platform environment', $choices, $choices[0]);
-    $chosenEnvironmentIndex = array_search($chosenEnvironmentLabel, $choices, TRUE);
-
-    return $applicationEnvironments[$chosenEnvironmentIndex];
-  }
-
-  /**
-   * @return array<mixed>
-   */
-  private function promptChooseDatabases(
-    EnvironmentResponse $cloudEnvironment,
-    DatabasesResponse $environmentDatabases,
-    bool $multipleDbs
-  ): array {
-    $choices = [];
-    if ($multipleDbs) {
-      $choices['all'] = 'All';
-    }
-    $defaultDatabaseIndex = 0;
-    if ($this->isAcsfEnv($cloudEnvironment)) {
-      $acsfSites = $this->getAcsfSites($cloudEnvironment);
-    }
-    foreach ($environmentDatabases as $index => $database) {
-      $suffix = '';
-      if (isset($acsfSites)) {
-        foreach ($acsfSites['sites'] as $domain => $acsfSite) {
-          if ($acsfSite['conf']['gardens_db_name'] === $database->name) {
-            $suffix .= ' (' . $domain . ')';
-            break;
-          }
-        }
-      }
-      if ($database->flags->default) {
-        $defaultDatabaseIndex = $index;
-        $suffix .= ' (default)';
-      }
-      $choices[] = $database->name . $suffix;
-    }
-
-    $question = new ChoiceQuestion(
-      $multipleDbs ? 'Choose databases. You may choose multiple. Use commas to separate choices.' : 'Choose a database.',
-      $choices,
-      $defaultDatabaseIndex
-    );
-    $question->setMultiselect($multipleDbs);
-    if ($multipleDbs) {
-      $chosenDatabaseKeys = $this->io->askQuestion($question);
-      $chosenDatabases = [];
-      if (count($chosenDatabaseKeys) === 1 && $chosenDatabaseKeys[0] === 'all') {
-        if (count($environmentDatabases) > 10) {
-          $this->io->warning('You have chosen to pull down more than 10 databases. This could exhaust your disk space.');
-        }
-        return (array) $environmentDatabases;
-      }
-      foreach ($chosenDatabaseKeys as $chosenDatabaseKey) {
-        $chosenDatabases[] = $environmentDatabases[$chosenDatabaseKey];
-      }
-
-      return $chosenDatabases;
-    }
-
-    $chosenDatabaseLabel = $this->io->choice('Choose a database', $choices, $defaultDatabaseIndex);
-    $chosenDatabaseIndex = array_search($chosenDatabaseLabel, $choices, TRUE);
-    return [$environmentDatabases[$chosenDatabaseIndex]];
-  }
-
-  protected function runComposerScripts(callable $outputCallback = NULL): void {
-    if (!file_exists(Path::join($this->dir, 'composer.json'))) {
-      $this->io->note('composer.json file not found. Skipping composer install.');
-      return;
-    }
-    if (!$this->localMachineHelper->commandExists('composer')) {
-      $this->io->note('Composer not found. Skipping composer install.');
-      return;
-    }
-    if (file_exists(Path::join($this->dir, 'vendor'))) {
-      $this->io->note('Composer dependencies already installed. Skipping composer install.');
-      return;
-    }
-    $this->checklist->addItem("Installing Composer dependencies");
-    $this->composerInstall($outputCallback);
-    $this->checklist->completePreviousItem();
-  }
-
   private function determineSite(string|EnvironmentResponse|array $environment, InputInterface $input): mixed {
     if (isset($this->site)) {
       return $this->site;
@@ -564,27 +441,6 @@ abstract class PullCommandBase extends CommandBase {
     return $site;
   }
 
-  protected function rsyncFiles(string $sourceDir, string $destinationDir, ?callable $outputCallback): void {
-    $this->localMachineHelper->checkRequiredBinariesExist(['rsync']);
-    $command = [
-      'rsync',
-      // -a archive mode; same as -rlptgoD.
-      // -z compress file data during the transfer.
-      // -v increase verbosity.
-      // -P show progress during transfer.
-      // -h output numbers in a human-readable format.
-      // -e specify the remote shell to use.
-      '-avPhze',
-      'ssh -o StrictHostKeyChecking=no',
-      $sourceDir . '/',
-      $destinationDir,
-    ];
-    $process = $this->localMachineHelper->execute($command, $outputCallback, NULL, ($this->output->getVerbosity() > OutputInterface::VERBOSITY_NORMAL));
-    if (!$process->isSuccessful()) {
-      throw new AcquiaCliException('Unable to sync files. {message}', ['message' => $process->getErrorOutput()]);
-    }
-  }
-
   private function rsyncFilesFromCloud(EnvironmentResponse $chosenEnvironment, Closure $outputCallback, string $site): void {
     $sourceDir = $chosenEnvironment->sshUrl . ':' . $this->getCloudFilesDir($chosenEnvironment, $site);
     $destinationDir = $this->getLocalFilesDir($site);
@@ -593,34 +449,7 @@ abstract class PullCommandBase extends CommandBase {
     $this->rsyncFiles($sourceDir, $destinationDir, $outputCallback);
   }
 
-  /**
-   * @param string|null $site
-   * @return DatabaseResponse[]
-   */
-  protected function determineCloudDatabases(Client $acquiaCloudClient, EnvironmentResponse $chosenEnvironment, string $site = NULL, bool $multipleDbs = FALSE): array {
-    $databasesRequest = new Databases($acquiaCloudClient);
-    $databases = $databasesRequest->getAll($chosenEnvironment->uuid);
-
-    if (count($databases) === 1) {
-      $this->logger->debug('Only a single database detected on Cloud');
-      return [$databases[0]];
-    }
-    $this->logger->debug('Multiple databases detected on Cloud');
-    if ($site && !$multipleDbs) {
-      if ($site === 'default') {
-        $this->logger->debug('Site is set to default. Assuming default database');
-        $site = self::getSiteGroupFromSshUrl($chosenEnvironment->sshUrl);
-      }
-      $databaseNames = array_column((array) $databases, 'name');
-      $databaseKey = array_search($site, $databaseNames, TRUE);
-      if ($databaseKey !== FALSE) {
-        return [$databases[$databaseKey]];
-      }
-    }
-    return $this->promptChooseDatabases($chosenEnvironment, $databases, $multipleDbs);
-  }
-
-  private function determineCloneProject(OutputInterface $output): bool {
+  protected function determineCloneProject(OutputInterface $output): bool {
     $finder = $this->localMachineHelper->getFinder()->files()->in($this->dir)->ignoreDotFiles(FALSE);
 
     // If we are in an IDE, assume we should pull into /home/ide/project.
@@ -665,29 +494,6 @@ abstract class PullCommandBase extends CommandBase {
     $this->projectDir = $this->dir;
   }
 
-  protected function determineEnvironment(InputInterface $input, OutputInterface $output, bool $allowProduction = FALSE): array|string|EnvironmentResponse {
-    if (isset($this->sourceEnvironment)) {
-      return $this->sourceEnvironment;
-    }
-
-    if ($input->getArgument('environmentId')) {
-      $environmentId = $input->getArgument('environmentId');
-      $chosenEnvironment = $this->getCloudEnvironment($environmentId);
-    }
-    else {
-      $cloudApplicationUuid = $this->determineCloudApplication();
-      $cloudApplication = $this->getCloudApplication($cloudApplicationUuid);
-      $output->writeln('Using Cloud Application <options=bold>' . $cloudApplication->name . '</>');
-      $acquiaCloudClient = $this->cloudApiClientService->getClient();
-      $chosenEnvironment = $this->promptChooseEnvironment($acquiaCloudClient, $cloudApplicationUuid, $allowProduction);
-    }
-    $this->logger->debug("Using environment $chosenEnvironment->label $chosenEnvironment->uuid");
-
-    $this->sourceEnvironment = $chosenEnvironment;
-
-    return $this->sourceEnvironment;
-  }
-
   protected function checkEnvironmentPhpVersions(EnvironmentResponse $environment): void {
     $version = $this->getIdePhpVersion();
     if (empty($version)) {
@@ -715,76 +521,6 @@ abstract class PullCommandBase extends CommandBase {
   private function environmentPhpVersionMatches(EnvironmentResponse $environment): bool {
     $currentPhpVersion = $this->getIdePhpVersion();
     return $environment->configuration->php->version === $currentPhpVersion;
-  }
-
-  protected function executeAllScripts(\Symfony\Component\Console\Input\InputInterface $input, Closure $outputCallback): void {
-    $this->setDirAndRequireProjectCwd($input);
-    $this->runComposerScripts($outputCallback);
-    $this->runDrushCacheClear($outputCallback);
-    $this->runDrushSqlSanitize($outputCallback);
-  }
-
-  protected function runDrushCacheClear(Closure $outputCallback): void {
-    if ($this->getDrushDatabaseConnectionStatus()) {
-      $this->checklist->addItem('Clearing Drupal caches via Drush');
-      // @todo Add support for Drush 8.
-      $process = $this->localMachineHelper->execute([
-        'drush',
-        'cache:rebuild',
-        '--yes',
-        '--no-interaction',
-        '--verbose',
-      ], $outputCallback, $this->dir, FALSE);
-      if (!$process->isSuccessful()) {
-        throw new AcquiaCliException('Unable to rebuild Drupal caches via Drush. {message}', ['message' => $process->getErrorOutput()]);
-      }
-      $this->checklist->completePreviousItem();
-    }
-    else {
-      $this->logger->notice('Drush does not have an active database connection. Skipping cache:rebuild');
-    }
-  }
-
-  protected function runDrushSqlSanitize(Closure $outputCallback): void {
-    if ($this->getDrushDatabaseConnectionStatus()) {
-      $this->checklist->addItem('Sanitizing database via Drush');
-      $process = $this->localMachineHelper->execute([
-        'drush',
-        'sql:sanitize',
-        '--yes',
-        '--no-interaction',
-        '--verbose',
-      ], $outputCallback, $this->dir, FALSE);
-      if (!$process->isSuccessful()) {
-        throw new AcquiaCliException('Unable to sanitize Drupal database via Drush. {message}', ['message' => $process->getErrorOutput()]);
-      }
-      $this->checklist->completePreviousItem();
-      $this->io->newLine();
-      $this->io->text('Your database was sanitized via <options=bold>drush sql:sanitize</>. This has changed all user passwords to randomly generated strings. To log in to your Drupal site, use <options=bold>drush uli</>');
-    }
-    else {
-      $this->logger->notice('Drush does not have an active database connection. Skipping sql:sanitize.');
-    }
-  }
-
-  private function composerInstall(?callable $outputCallback): void {
-    $process = $this->localMachineHelper->execute([
-      'composer',
-      'install',
-      '--no-interaction',
-    ], $outputCallback, $this->dir, FALSE);
-    if (!$process->isSuccessful()) {
-      throw new AcquiaCliException('Unable to install Drupal dependencies via Composer. {message}',
-        ['message' => $process->getErrorOutput()]);
-    }
-  }
-
-  protected function getHostFromDatabaseResponse(mixed $environment, DatabaseResponse $database): string {
-    if ($this->isAcsfEnv($environment)) {
-      return $database->db_host . '.enterprise-g1.hosting.acquia.com';
-    }
-
-    return $database->db_host;
   }
 
   private function getDatabaseBackup(
