+++ src/Command/Pull/PullFilesCommand.php
@@ -22,7 +22,9 @@ final class PullFilesCommand extends PullCommandBase {
 
   protected function execute(InputInterface $input, OutputInterface $output): int {
     parent::execute($input, $output);
-    $this->pullFiles($input, $output);
+    $this->setDirAndRequireProjectCwd($input);
+    $sourceEnvironment = $this->determineEnvironment($input, $output, TRUE);
+    $this->pullFiles($input, $output, $sourceEnvironment);
 
     return Command::SUCCESS;
   }
