+++ src/Command/Pull/PullScriptsCommand.php
@@ -4,6 +4,8 @@ declare(strict_types = 1);
 
 namespace Acquia\Cli\Command\Pull;
 
+use Acquia\Cli\Command\CommandBase;
+use Acquia\Cli\Output\Checklist;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
@@ -11,7 +13,9 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
 #[AsCommand(name: 'pull:run-scripts', description: 'Execute post pull scripts')]
-final class PullScriptsCommand extends PullCommandBase {
+final class PullScriptsCommand extends CommandBase {
+
+  protected Checklist $checklist;
 
   protected function configure(): void {
     $this
@@ -19,8 +23,14 @@ final class PullScriptsCommand extends PullCommandBase {
       ->addOption('dir', NULL, InputArgument::OPTIONAL, 'The directory containing the Drupal project to be refreshed');
   }
 
+  protected function initialize(InputInterface $input, OutputInterface $output): void {
+    parent::initialize($input, $output);
+    $this->checklist = new Checklist($output);
+  }
+
   protected function execute(InputInterface $input, OutputInterface $output): int {
-    $this->executeAllScripts($input, $this->getOutputCallback($output, $this->checklist));
+    $this->setDirAndRequireProjectCwd($input);
+    $this->executeAllScripts($this->getOutputCallback($output, $this->checklist), $this->checklist);
 
     return Command::SUCCESS;
   }
