+++ src/Command/Push/PushCodeCommand.php
@@ -5,7 +5,6 @@ declare(strict_types = 1);
 namespace Acquia\Cli\Command\Push;
 
 use Acquia\Cli\Attribute\RequireAuth;
-use Acquia\Cli\Command\Pull\PullCommandBase;
 use Acquia\DrupalEnvironmentDetector\AcquiaDrupalEnvironmentDetector;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
@@ -14,7 +13,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 #[RequireAuth]
 #[AsCommand(name: 'push:code', description: 'Push code from your IDE to a Cloud Platform environment')]
-final class PushCodeCommand extends PullCommandBase {
+final class PushCodeCommand extends PushCommandBase {
 
   protected function configure(): void {
     $this
