+++ tests/phpunit/src/Commands/Acsf/AcsfApiCommandTest.php
@@ -141,7 +141,6 @@ class AcsfApiCommandTest extends AcsfCommandTestBase {
       $this->sshHelper,
       $this->sshDir,
       $this->logger,
-      $this->httpClientProphecy->reveal()
     );
   }
