+++ tests/phpunit/src/Commands/Pull/PullDatabaseCommandTest.php
@@ -11,6 +11,7 @@ use Acquia\Cli\Exception\AcquiaCliException;
 use Acquia\Cli\Helpers\LocalMachineHelper;
 use Acquia\Cli\Helpers\SshHelper;
 use AcquiaCloudApi\Response\BackupsResponse;
+use GuzzleHttp\Client;
 use GuzzleHttp\Exception\RequestException;
 use GuzzleHttp\Psr7\Uri;
 use Prophecy\Argument;
@@ -58,7 +59,21 @@ class PullDatabaseCommandTest extends PullCommandTestBase {
   }
 
   protected function createCommand(): CommandBase {
-    return $this->injectCommand(PullDatabaseCommand::class);
+    $this->httpClientProphecy = $this->prophet->prophesize(Client::class);
+
+    return new PullDatabaseCommand(
+      $this->localMachineHelper,
+      $this->datastoreCloud,
+      $this->datastoreAcli,
+      $this->cloudCredentials,
+      $this->telemetryHelper,
+      $this->acliRepoRoot,
+      $this->clientServiceProphecy->reveal(),
+      $this->sshHelper,
+      $this->sshDir,
+      $this->logger,
+      $this->httpClientProphecy->reveal()
+    );
   }
 
   public function testPullDatabases(): void {
